+++ src/Entity/Html/Link.php
@@ -3,10 +3,22 @@
 namespace VysokeSkoly\UtilsBundle\Entity\Html;
 
 use Assert\Assertion;
+use MF\Collection\Immutable\Generic\KVPair;
+use MF\Collection\Immutable\Generic\ListCollection;
+use MF\Collection\Immutable\Generic\Map;
 
 class Link
 {
     private array $parameters;
+
+    public static function fromDomElement(\DOMElement $element): self
+    {
+        $parameters = Map::fromPairs(
+            ListCollection::create($element->attributes, fn (\DOMAttr $attr) => new KVPair($attr->name, $attr->value)),
+        );
+
+        return new self($parameters->toArray());
+    }
 
     public function __construct(array $parameters)
     {